feat: add new test for config generation
